# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArnieAI Vulnerability Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServiceSummary {
  status: string;
  total: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
}

export interface ArtifactFolder {
  name: string;
  count: number;
  /** @nullable */
  lastModified: string | null;
}

export interface S3Summary {
  bucket: string;
  folders: ArtifactFolder[];
}

export interface ActivityItem {
  id: string;
  source: string;
  title: string;
  severity: string;
  timestamp: string;
}

export interface DashboardSummary {
  guardduty: ServiceSummary;
  inspector: ServiceSummary;
  securityhub: ServiceSummary;
  s3: S3Summary;
  recentActivity: ActivityItem[];
}

export interface ArtifactObject {
  key: string;
  folder: string;
  size: number;
  lastModified: string;
  /** @nullable */
  etag: string | null;
}

export interface PresignedUrl {
  url: string;
  expiresIn: number;
}

export interface LogGroup {
  name: string;
  /** @nullable */
  retentionDays: number | null;
  /** @nullable */
  storedBytes: number | null;
}

export interface LogEvent {
  timestamp: number;
  message: string;
  /** @nullable */
  ingestionTime: number | null;
}

export interface GuardDutyFinding {
  id: string;
  title: string;
  description: string;
  severity: number;
  severityLabel: string;
  type: string;
  region: string;
  accountId: string;
  /** @nullable */
  resourceType: string | null;
  createdAt: string;
  updatedAt: string;
  count: number;
}

export interface InspectorFinding {
  arn: string;
  title: string;
  description: string;
  severity: string;
  status: string;
  type: string;
  /** @nullable */
  resourceType: string | null;
  /** @nullable */
  resourceId: string | null;
  /** @nullable */
  cveId: string | null;
  /** @nullable */
  cvssScore: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface SecurityHubFinding {
  id: string;
  title: string;
  description: string;
  severity: string;
  severityScore: number;
  status: string;
  /** @nullable */
  productName: string | null;
  /** @nullable */
  resourceType: string | null;
  /** @nullable */
  resourceId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AgentSession {
  id: number;
  name: string;
  status: string;
  /** @nullable */
  conversationId: number | null;
  scopeInTargets: string;
  scopeOutTargets: string;
  conditions: string;
  objectives: string;
  createdAt: string;
  updatedAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ToolInvocationRecord {
  id: number;
  sessionId: number;
  toolName: string;
  toolInput: string;
  toolOutput: string;
  scopeAllowed: boolean;
  scopeReason: string;
  createdAt: string;
}

export interface AgentSessionDetail {
  id: number;
  name: string;
  status: string;
  /** @nullable */
  conversationId: number | null;
  scopeInTargets: string;
  scopeOutTargets: string;
  conditions: string;
  objectives: string;
  createdAt: string;
  updatedAt: string;
  messages: AnthropicMessage[];
  toolInvocations: ToolInvocationRecord[];
}

export interface CreateAgentSessionBody {
  name: string;
  scopeInTargets: string[];
  scopeOutTargets?: string[];
  conditions?: string;
  objectives: string;
}

export interface SendAgentMessageBody {
  content: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface VirusTotalResult {
  ioc: string;
  iocType: string;
  malicious: number;
  suspicious: number;
  undetected: number;
  harmless: number;
  total: number;
  /** @nullable */
  reputation: number | null;
  /** @nullable */
  lastAnalysisDate: string | null;
  tags: string[];
  /** @nullable */
  permalink: string | null;
}

export type ListGuardDutyFindingsParams = {
  severity?: string;
  limit?: number;
};

export type ListInspectorFindingsParams = {
  severity?: string;
  limit?: number;
};

export type ListSecurityHubFindingsParams = {
  severity?: string;
  limit?: number;
};

export type GetLogEventsParams = {
  group: string;
  limit?: number;
  startTime?: number;
};

export type VirusTotalLookupParams = {
  /**
   * Hash (MD5/SHA1/SHA256), IP address, or domain to look up
   */
  ioc: string;
};