# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArnieAI Vulnerability Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agent/sessions:
    get:
      operationId: listAgentSessions
      tags: [agent]
      summary: List all agent sessions
      responses:
        "200":
          description: List of agent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSession"
    post:
      operationId: createAgentSession
      tags: [agent]
      summary: Create a new agent session with scope definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentSessionBody"
      responses:
        "201":
          description: Created agent session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSession"
  /agent/sessions/{id}:
    get:
      operationId: getAgentSession
      tags: [agent]
      summary: Get an agent session with messages and tool invocations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteAgentSession
      tags: [agent]
      summary: Delete an agent session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /agent/sessions/{id}/run:
    post:
      operationId: runAgentSession
      tags: [agent]
      summary: Start autonomous agent run (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of agent thoughts, tool calls, and results
          content:
            text/event-stream: {}
        "404":
          description: Not found
  /agent/sessions/{id}/message:
    post:
      operationId: sendAgentMessage
      tags: [agent]
      summary: Send a message to the agent (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAgentMessageBody"
      responses:
        "200":
          description: SSE stream of agent response
          content:
            text/event-stream: {}
        "404":
          description: Not found
  /agent/sessions/{id}/tool-invocations:
    get:
      operationId: listToolInvocations
      tags: [agent]
      summary: List tool invocations for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tool invocation log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolInvocationRecord"
...
components:
  schemas:
...
    AgentSession:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        conversationId:
          type: ["integer", "null"]
        scopeInTargets:
          type: string
        scopeOutTargets:
          type: string
        conditions:
          type: string
        objectives:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - conversationId
        - scopeInTargets
        - scopeOutTargets
        - conditions
        - objectives
        - createdAt
        - updatedAt
    AgentSessionDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
        conversationId:
          type: ["integer", "null"]
        scopeInTargets:
          type: string
        scopeOutTargets:
          type: string
        conditions:
          type: string
        objectives:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
        toolInvocations:
          type: array
          items:
            $ref: "#/components/schemas/ToolInvocationRecord"
      required:
        - id
        - name
        - status
        - conversationId
        - scopeInTargets
        - scopeOutTargets
        - conditions
        - objectives
        - createdAt
        - updatedAt
        - messages
        - toolInvocations
    CreateAgentSessionBody:
      type: object
      properties:
        name:
          type: string
        scopeInTargets:
          type: array
          items:
            type: string
        scopeOutTargets:
          type: array
          items:
            type: string
        conditions:
          type: string
        objectives:
          type: string
      required:
        - name
        - scopeInTargets
        - objectives
    SendAgentMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    ToolInvocationRecord:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        toolName:
          type: string
        toolInput:
          type: string
        toolOutput:
          type: string
        scopeAllowed:
          type: boolean
        scopeReason:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - toolName
        - toolInput
        - toolOutput
        - scopeAllowed
        - scopeReason
        - createdAt
...
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
...