# lib/api-zod/src/generated/types/agentSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArnieAI Vulnerability Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentSession {
  id: number;
  name: string;
  status: string;
  /** @nullable */
  conversationId: number | null;
  scopeInTargets: string;
  scopeOutTargets: string;
  conditions: string;
  objectives: string;
  createdAt: string;
  updatedAt: string;
}