# lib/api-zod/src/generated/types/createAgentSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArnieAI Vulnerability Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAgentSessionBody {
  name: string;
  scopeInTargets: string[];
  scopeOutTargets?: string[];
  conditions?: string;
  objectives: string;
}