# lib/api-zod/src/generated/types/sendAgentMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArnieAI Vulnerability Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SendAgentMessageBody {
  content: string;
}