# lib/api-zod/src/generated/types/toolInvocationRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArnieAI Vulnerability Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ToolInvocationRecord {
  id: number;
  sessionId: number;
  toolName: string;
  toolInput: string;
  toolOutput: string;
  scopeAllowed: boolean;
  scopeReason: string;
  createdAt: string;
}